redis and db healthy

+++ api/models.py
@@ -1,11 +1,11 @@
-from sqlalchemy import Column, Integer, String, DateTime, Boolean
+from sqlalchemy import Column, Integer, String, DateTime, Boolean, BigInteger
 from sqlalchemy.sql import func
 from api.database import Base
 
 class Url(Base):
     __tablename__ = "urls"
     
-    id = Column(Integer, primary_key=True, autoincrement=True)
+    id = Column(BigInteger, primary_key=True, autoincrement=True)
     original_url = Column(String, nullable=False)
     custom_alias = Column(String, unique=True, nullable=True)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
@@ -21,8 +21,8 @@ class Url(Base):
 class Click(Base):
     __tablename__ = "clicks"
     
-    id = Column(Integer, primary_key=True, autoincrement=True)
-    url_id = Column(Integer, nullable=False)  # Foreign key to urls.id
+    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    url_id = Column(BigInteger, nullable=False)  # Foreign key to urls.id
     ip_address = Column(String, nullable=True)
     user_agent = Column(String, nullable=True)
     referer = Column(String, nullable=True)
